Leaderboard improvements
Change weekleaderboard, monthleaderboard and historical leaderboard to weektopgainers, monthtopgainers and top gainers respectively.
Add comments to gettopgainers and update leaderboard
Add return current_user.balance to leaderboard route
Change totalvalue to DECIMAL(20, 2) instead of float in model.py due to issues with float using large numbers.

+++ models.py
@@ -177,7 +177,7 @@ class Leaderboard(Base):
                       primary_key=True)
     recordtime = Column(DateTime, primary_key=True)
     ranking = Column(Integer, nullable=False, unique=False)
-    totalvalue = Column(Float, unique=False, nullable=False)
+    totalvalue = Column(DECIMAL(20, 2), unique=False, nullable=False)
 
 # Allow creation of tables by running API directly
 if __name__ == "__main__":
